+++ MessengerProj/src/main/java/com/b44t/ui/SettingsNameActivity.java
@@ -261,7 +261,6 @@ public class SettingsNameActivity extends BaseFragment {
                     ((TextInfoCell) view).setText(mContext.getString(R.string.MyStatusExplain));
                     view.setBackgroundResource(R.drawable.greydivider_bottom);
                 }
-
             }
             return view;
         }